Bug fix for app crash
