+++ collect_app/src/main/java/org/odk/collect/android/activities/CaptureSelfieActivity.java
@@ -134,4 +134,17 @@ public class CaptureSelfieActivity extends Activity {
             });
         }
     }
+
+    public static boolean isFrontCameraAvailable() {
+        //https://developer.android.com/guide/topics/media/camera.html#check-camera-features
+        for (int camNo = 0; camNo < Camera.getNumberOfCameras(); camNo++) {
+            Camera.CameraInfo camInfo = new Camera.CameraInfo();
+            Camera.getCameraInfo(camNo, camInfo);
+            if (camInfo.facing == Camera.CameraInfo.CAMERA_FACING_FRONT) {
+
+                return true;
+            }
+        }
+        return false; // No front-facing camera found
+    }
 }
